SwordMaster I
This is the first iteration of a game I'm working on called SwordMaster. The plot follows a retro-styled cat sword-fighter. I'm looking forward to pushing future versions to github to monitor my own progress.

The Javascript code for the game is located in #cat_runner.js#
A zipped version of the code, in addition to all of the .png and .mp3 files, is included as #interactive 5.zip#

Check it out if you have a chance!